Fixed bug: (signed) integer overflow

+++ crypto_sign/qtesla-p-I/clean/poly.c
@@ -17,7 +17,7 @@ static int64_t reduce(int64_t a) {
     // Montgomery reduction
     int64_t u;
 
-    u = (a * PARAM_QINV) & 0xFFFFFFFF;
+    u = ((uint64_t)a * PARAM_QINV) & 0xFFFFFFFF;
     u *= PARAM_Q;
     a += u;
     return a >> 32;
